requirements.txt
ai2thor==3.1.0
boto3==1.24.27
fuzzywuzzy==0.18.0
networkx==2.5
pydub==0.24.1
python-Levenshtein
tqdm
pillow
pydantic==1.8.2
numpy
pandas
opencv-python
vocab
revtok
sacred
etaprogress
scikit-video
lmdb
gtimer
filelock
termcolor
tensorboardX==1.8
Flask==2.0.3
flask_restful==0.3.9
Werkzeug==2.2.3
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
flask_wtf==1.0.1
markupsafe==2.1.1
future
pandoc
six
typing
ConfigArgParse
protobuf==3.20.2
h5py
Pillow
matplotlib
tensorboard
seaborn==0.9.0
imageio==2.6.0
scikit-fmm
scikit-image==0.15.0
scikit-learn==0.22.2.post1
ifcfg
ipdb
transformers
numpy-quaternion
tiktoken
moviepy
wandb
timm==0.6.7
openai